Adding 'safe' group spawning. Needs massive testing.
