+++ addons/core/XEH_PREP.hpp
@@ -1,3 +1,6 @@
 TRACE_1("",QUOTE(ADDON));
 
+PREP(createGroup);
 PREP(isAuthorized);
+PREP(canCreateGroup);
+PREP(sendCuratorHint);
