+++ addons/zeusHC/functions/fnc_buildAGroupDialogSpawn.sqf
@@ -5,14 +5,17 @@ TRACE_1("Params",_this);
 GVAR(lastGroupFactionIndex) = lbCurSel BUILD_GROUP_FACTIONS_IDC;
 
 // get items in the to spawn bin
+private _unitsToSave = [];
 private _unitsToAdd = [];
 for "_index" from 0 to ((lbSize BUILD_GROUP_TO_ADD_IDC) - 1) do {
-    _unitsToAdd pushBack [lbData [BUILD_GROUP_TO_ADD_IDC, _index],
-                          lbText [BUILD_GROUP_TO_ADD_IDC, _index],
-                          lbPicture [BUILD_GROUP_TO_ADD_IDC, _index]];
+    _unitsToSave pushBack [lbData [BUILD_GROUP_TO_ADD_IDC, _index],
+                           lbText [BUILD_GROUP_TO_ADD_IDC, _index],
+                           lbPicture [BUILD_GROUP_TO_ADD_IDC, _index]];
+
+    _unitsToAdd pushBack lbData [BUILD_GROUP_TO_ADD_IDC, _index;
 };
 
-GVAR(lastGroupUnitArray) = _unitsToAdd;
+GVAR(lastGroupUnitArray) = _unitsToSave;
 [GVAR(buildSpawnLocation),_unitsToAdd] remoteExecCall [QFUNC(buildAGroupSpawnServer), SERVER_CLIENT_ID];
 
 // last thing, close dialog with success
